+++ locksmith/src/controllers/v2/priceController.ts
@@ -1,17 +1,13 @@
 import { RequestHandler } from 'express'
-
-import {
-  createPricingForPurchase,
-  createTotalCharges,
-  defiLammaPrice,
-} from '../../utils/pricing'
+import * as priceOperations from '../../operations/pricingOperations'
+import { createPricingForPurchase } from '../../utils/pricing'
 import { ethers } from 'ethers'
 import { getCreditCardEnabledStatus } from '../../operations/creditCardOperations'
 import * as Normalizer from '../../utils/normalizer'
 import { Web3Service } from '@unlock-protocol/unlock-js'
 import networks from '@unlock-protocol/networks'
+import { MIN_PAYMENT_STRIPE } from '../../utils/constants'
 
-const MIN_PAYMENT_STRIPE = 100
 export const amount: RequestHandler = async (request, response) => {
   const network = Number(request.params.network || 1)
   const amount = parseFloat(request.query.amount?.toString() || '1')
@@ -20,7 +16,7 @@ export const amount: RequestHandler = async (request, response) => {
     ? erc20Address
     : undefined
 
-  const result = await defiLammaPrice({
+  const result = await priceOperations.getUsdPricingForLock({
     network,
     amount,
     address,
@@ -38,9 +34,9 @@ export const total: RequestHandler = async (request, response) => {
     ? erc20Address
     : undefined
 
-  const charge = await createTotalCharges({
-    network,
+  const charge = await priceOperations.getTotalCharges({
     amount,
+    network,
     address,
   })
 
@@ -110,7 +106,7 @@ export const isCardPaymentEnabledForLock: RequestHandler = async (
   const web3Service = new Web3Service(networks)
   const lock = await web3Service.getLock(lockAddress, network)
 
-  const result = await defiLammaPrice({
+  const result = await priceOperations.getDefiLammaPrice({
     network,
     address: lock?.currencyContractAddress,
     amount: Number(`${lock.keyPrice}`),
